Add bb_asg_lifecycle_hook
In the following commit, we removed logic from bb_scheduler to talk to
AWS ASG lifecycle hooks:

https://github.com/buildbarn/bb-remote-execution/commit/78e027fca8f92f0f43a7eb6052f16ece682f8674

There is no need to let this be part of bb_scheduler itself, as we can
easily move that into a separate binary. This change adds such a binary
to the bb-autoscaler repository. Even if people don't need this binary
in literal form, it still acts as useful documentation for how to add
graceful worker termination to your cluster.